Convert CheckoutMasterbar to func component & use useValidCheckoutBackUrl hook.
