+++ client/layout/masterbar/logged-out.jsx
@@ -148,7 +148,7 @@ class MasterbarLoggedOut extends React.Component {
 		if ( isCheckout ) {
 			return (
 				<AsyncLoad
-					require="calypso/layout/masterbar/checkout"
+					require="calypso/layout/masterbar/checkout.tsx"
 					placeholder={ null }
 					title={ title }
 				/>
